Updates for Flask and SqlAlchemy libraries
