+++ src/server/FlaskSqlObjs.py
@@ -1,0 +1,10 @@
+# Helper module to host references to Flask APP and SQLAlchemy database
+
+APP = None
+DB = None
+
+def set_objects(app_obj, db_obj):
+    global APP
+    APP = app_obj
+    global DB
+    DB = db_obj
