+++ src/server/plugins/rh_actions_builtin/__init__.py
@@ -1,10 +1,12 @@
 ''' builtin Actions '''
 
-import json
-import RHUtils
+import RHData
 from eventmanager import Evt
 from EventActions import ActionEffect
 from RHUI import UIField, UIFieldType, UIFieldSelectOption
+
+from FlaskSqlObjs import APP
+APP.app_context().push()
 
 class ActionsBuiltin():
     def __init__(self, rhapi):
@@ -12,24 +14,24 @@ class ActionsBuiltin():
 
     def speakEffect(self, action, args):
         if 'text' in action:
-            text = RHUtils.doReplace(self._rhapi, action['text'], args, True)
+            text = RHData.doReplace(self._rhapi, action['text'], args, True)
             self._rhapi.ui.message_speak(text)
 
     def messageEffect(self, action, args):
         if 'text' in action:
-            text = RHUtils.doReplace(self._rhapi, action['text'], args)
+            text = RHData.doReplace(self._rhapi, action['text'], args)
             self._rhapi.ui.message_notify(text)
 
     def alertEffect(self, action, args):
         if 'text' in action:
-            text = RHUtils.doReplace(self._rhapi, action['text'], args)
+            text = RHData.doReplace(self._rhapi, action['text'], args)
             self._rhapi.ui.message_alert(text)
 
 
     def clearMessagesEffect(self, action, args):
         self._rhapi.ui.clear_messages()
         if 'text' in action:
-            text = RHUtils.doReplace(self._rhapi, action['text'], args)
+            text = RHData.doReplace(self._rhapi, action['text'], args)
             self._rhapi.ui.message_notify(text)
 
     def scheduleEffect(self, action, _args):
