+++ src/server/requirements.txt
@@ -1,9 +1,9 @@
 # Flask
-Flask==1.1.*
-itsdangerous==1.1.*
-Jinja2==2.11.*
-Werkzeug==1.0.*
-MarkupSafe==1.1.*
+Flask==3.0.*
+itsdangerous==2.1.*
+Jinja2==3.1.*
+Werkzeug==3.0.*
+MarkupSafe==2.1.*
 
 # Socket
 Flask-SocketIO==5.3.*
@@ -12,8 +12,8 @@ python-engineio==4.4.*
 websocket-client==1.5.*
 
 # Database
-flask_sqlalchemy==2.4.*
-SQLAlchemy==1.3.*
+flask_sqlalchemy==3.1.*
+SQLAlchemy==2.0.*
 
 # gevent
 gevent==23.9.*
